requirements.txt
fastapi
uvicorn
slack_sdk
langchain
google-generativeai
requests
dotenv
aiohttp
jira
atlassian-python-api